refactor helper functions to helper and another printer to output
remove unused imports
